feat: admin 권한 전체 board update, delete가능하게 구현

+++ src/main/java/week13/board/service/BoardService.java
@@ -2,6 +2,7 @@ package week13.board.service;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
@@ -13,6 +14,7 @@ import week13.board.exception.ErrorCode;
 import week13.board.repository.BoardRepository;
 import week13.board.repository.UserRepository;
 
+import java.util.Collection;
 import java.util.List;
 
 @Service
@@ -55,10 +57,7 @@ public class BoardService {
         Post post = boardRepository.findById(id)
                 .orElseThrow(() -> new CustomException(ErrorCode.POST_NOT_FOUND));
 
-        String currentUsername = getCurrentUsername();
-        if (!post.getUser().getUsername().equals(currentUsername)) {
-            throw new CustomException(ErrorCode.UNAUTHORIZED_USER);
-        }
+        verifyPermission(post);
         post.update(requestDto);
 
         return post;
@@ -66,6 +65,30 @@ public class BoardService {
 
     @Transactional
     public void deletePost(Long id) {
+        Post post = boardRepository.findById(id)
+                .orElseThrow(() -> new CustomException(ErrorCode.POST_NOT_FOUND));
+
+        verifyPermission(post);
         boardRepository.deleteById(id);
+    }
+
+    private void verifyPermission(Post post) {
+        String currentUsername = getCurrentUsername();
+        if (!post.getUser().getUsername().equals(currentUsername) && !isAdmin()) {
+            throw new CustomException(ErrorCode.UNAUTHORIZED_USER);
+        }
+    }
+
+    // 현재 사용자가 관리자 권한을 가지고 있는지 확인하는 메서드
+    private boolean isAdmin() {
+        Collection<? extends GrantedAuthority> authorities =
+                SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+
+        for (GrantedAuthority authority : authorities) {
+            if (authority.getAuthority().equals("ADMIN")) {
+                return true;
+            }
+        }
+        return false;
     }
 }
